Stage 2b: converge 7 of 9 module pairs into the canonical modules (#1188)
Fold each X4e's genuinely-new 4e algorithms into the canonical aima/X.py, keep
the used base for shared symbols, then delete the X4e duplicate + reconcile tests:
- games: + monte_carlo_tree_search, mcts_player (MCTS); games4e removed
- mdp: + pomdp_lookahead, update_belief, q_value (DDN); mdp4e removed
- nlp: + Tree, subspan, TextParsingProblem, astar_search_parsing, beam_search_parsing; nlp4e removed
- probability: + is_independent, gen_possible_events, gaussian_probability,
  logistic_probability, ContinuousBayesNode, complied_burglary; probability4e removed
- utils: + gaussian_kernel(_1D/_2D), conv1D, map_vector, MCT_Node, ucb (used by
  the MCTS above); perception4e and making_simple_decisions4e repointed to aima.utils
Notebooks and docs updated to the canonical modules. Skipped naming-variant
duplicates (expect_min_max_player, model_selection, KB_AgentProgram, extend).

DEFERRED: the utils/learning convergence for the 4e ML cluster (utils4e,
learning4e, deep_learning4e) — they are mutually coupled through the 4e
DataSet.sanitize convention (drops the target) and the recursive vector helpers,
so unifying them requires porting that cluster as its own effort. utils4e and
learning4e are kept for now.

Full suite: 407 passed; docs build with 0 warnings.
