+++ docs/probability.rst
@@ -6,17 +6,7 @@ Probability, MDPs & Decisions
    :undoc-members:
    :show-inheritance:
 
-.. automodule:: aima.probability4e
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 .. automodule:: aima.mdp
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-.. automodule:: aima.mdp4e
    :members:
    :undoc-members:
    :show-inheritance:
