+++ docs/utils.rst
@@ -6,11 +6,6 @@ Utilities & Notebook Helpers
    :undoc-members:
    :show-inheritance:
 
-.. automodule:: aima.utils4e
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 .. automodule:: aima.notebook_utils
    :members:
    :undoc-members:
